Factored event_writer, file_event_writer, and sqlite_event_writer out of event_history into their own files for more modularity.
