+++ src/common/date_utils.hpp
@@ -18,6 +18,16 @@ public:
   static std::string currentDate();
 
   /**
+   * Get the current time in microseconds.
+   */
+  static long currentDateTimeInMicro();
+
+   /**
+    * Get a human readable timestamp in microseconds.
+    */
+  static std::string humanReadableDateTimeInMicro();
+
+  /**
    * Unit test utility method that makes this class return a fixed string
    * as the date instead of looking up the current time.
    */
